+++ DatabaseManager.h
@@ -24,13 +24,13 @@ class DatabaseManager {
 
 public:
   static int64_t registerNewScop(uint64_t hashID, const char *name);
-  static int64_t setScopParams(int64_t generalInfoID, int maxLoopDepth,
-                               int instructionNumber, int memoryAccess,
-                               int readMemoryAccess, int indvarNumber,
-                               int sumOfCoeff, int sumOfOffsets,
+  static int64_t setScopParams(int64_t generalInfoID, int64_t maxLoopDepth,
+                               int64_t instructionNumber, int64_t memoryAccess,
+                               int64_t readMemoryAccess, int64_t indvarNumber,
+                               int64_t sumOfCoeff, int64_t sumOfOffsets,
                                uint64_t *upperPartScopID,
                                uint64_t *lowerPartScopID);
   static int64_t setScopLoopParams(uint64_t upperPartScopID,
-                                   uint64_t lowerPartScopID, int range);
+                                   uint64_t lowerPartScopID, int64_t range);
 };
 #endif
