+++ DatabaseProxy.cpp
@@ -23,19 +23,20 @@ extern "C" int64_t register_new_scop(uint64_t hashID, const char *scopName) {
   return DatabaseManager::registerNewScop(hashID, scopName);
 }
 
-extern "C" int64_t set_scop_params(int64_t generalInfoID, int maxLoopDepth,
-                                   int instructionNumber, int memoryAccess,
-                                   int readMemoryAccess, int indvarNumber,
-                                   int sumOfCoeff, int sumOfOffsets,
-                                   uint64_t *upperPartScopID,
-                                   uint64_t *lowerPartScopID) {
+extern "C" int64_t
+set_scop_params(int64_t generalInfoID, int64_t maxLoopDepth,
+                int64_t instructionNumber, int64_t memoryAccess,
+                int64_t readMemoryAccess, int64_t indvarNumber,
+                int64_t sumOfCoeff, int64_t sumOfOffsets,
+                uint64_t *upperPartScopID, uint64_t *lowerPartScopID) {
   return DatabaseManager::setScopParams(
       generalInfoID, maxLoopDepth, instructionNumber, memoryAccess,
       readMemoryAccess, indvarNumber, sumOfCoeff, sumOfOffsets, upperPartScopID,
       lowerPartScopID);
 }
 extern "C" int64_t set_scop_loop_params(uint64_t upperPartScopID,
-                                        uint64_t lowerPartScopID, int range) {
+                                        uint64_t lowerPartScopID,
+                                        int64_t range) {
   return DatabaseManager::setScopLoopParams(upperPartScopID, lowerPartScopID,
                                             range);
 }
